# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API to manage a vehicle mechanic shop"
  version: "1.0.0"
host: "mechanic-shop-api-1kl1.onrender.com"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers/login: 
    post: 
      tags:
      - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customer credentials, generates and returns a customer specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse" 
          examples:
            application/json:
            token: "nvjdsnjngjjbsnjngsjnjsnbjvbsbjbgsj..."
            message: "login successful"
            status: "success"

        400:
          description: "Invalid email or password"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "invalid email or password"
            status: "Error"

  # create customer
  /customers/:
    post:
      tags:
        - Customers
      summary: "Create new customer endpoint"
      description: "Create a new customer in database, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          descriptions: "Successfull Customer Creation"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              phone: "1112223333"
              email: "johnd@email.com"
              password: "password123"

        400:
          description: "Missing data"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "missing data for required field."
            status: "Error"

  # get all customers
...
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Get customer by id endpoint"
      description: "Takes a customer id and returns the matching customer data."
      responses:
        200:
          description: "Retrieved Customer Successfully"           
          schema:
            $ref: "#/definitions/GetCustomerByID" 
          examples:
            application/json:
              id: 1
              name: "John Doe"
              phone: "1112223333"
              email: "johnd@email.com"
              password: "password123"

        404:
          description: "Customer not found"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "customer not found"
            status: "Error"

  # update customer by id
    put: 
      tags:
        - Customers
      summary: "Update customer endpoint, (requires login)"
      description: "Endpoint to update a customer. Token authenticated route, customer must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          descriptions: "Successfully Updated Customer"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "Dohn Joe"
              phone: "3335554444"
              email: "dohnj@email.com"
              password: "password123" 

        404:
          description: "Customer not found"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "customer not found"
            status: "Error"

        401:
          description: "Missing token"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "missing token"
            status: "Error"

        401:
          description: "Token expired"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "token expired"
            status: "Error"

        401:
          description: "Invalid token"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "invalid token"
            status: "Error"

        401:
          description: "Unauthorized"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "you must be logged in to access this."
            status: "Error"

  # delete customer by id
    delete:
      tags:
        - Customers
      summary: "Delete customer endpoint, (requires login)"
      description: "Endpoint to delete a customer. Token authenticated route, customer must be logged in to delete their account"
      security:
       - bearerAuth: []
      responses:
        200:
          description: "Successfully Deleted Customer"           
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "successfully deleted user <customer_id>."

        401:
          description: "Missing token"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "missing token"
            status: "Error"

        401:
          description: "Token expired"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "token expired"
            status: "Error"

        401:
          description: "Invalid token"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "invalid token"
            status: "Error"

        401:
          description: "Unauthorized"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "you must be logged in to access this."
            status: "Error"

# Mechanics
  # create mechanic
...
  /service_tickets/:
    post:
      tags:
        - ServiceTickets
      summary: "Create service ticket endpoint"
      description: "Creates a new service ticket."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new service ticket."
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        201:
          descriptions: "Successfull Service Ticket Creation"
          schema:
            $ref: "#/definitions/CreateServiceTicketResponse"
          examples:
            application/json:
              id: 1
              VIN: "12345678901234567"
              service_date: "2025-01-30"
              service_desc: "tires"
              customer_id: 1

        400:
          description: "Missing data"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "missing data for required field."
            status: "Error"

  # get all service tickets
...
  /service_tickets/{customer_id}:
    get:
      tags:
        - ServiceTickets
      summary: "Get service tickets by customer id endpoint"
      description: "Takes a customer id and returns the matching service tickets data associated with the customer."
      responses:
        200:
          description: "Retrieved service tickets successfully"           
          schema:
            $ref: "#/definitions/GetCustomerServiceTickets"

        404:
          description: "No customer service tickets"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "no customer service tickets found"
            status: "Error"

  # edit mechanics on service ticket
  /service_tickets/{ticket_id}/edit:
    put: 
      tags:
        - ServiceTickets
      summary: "Add/remove mechanic from a service ticket endpoint"
      description: "Takes a service ticket id to add or remove a mechanic(s) from the service ticket."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to add/remove a mechanic from a service ticket."
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicsTicketPayload"
      responses:
        200:
          descriptions: "Successfully update mechanics on service ticket"
          schema:
            $ref: "#/definitions/UpdateMechanicsTicketResponse"
          examples:
            application/json:
              ticket_id: 1
              add_mechanic_ids: [3]
              remove_mechanic_ids: [2]

        404:
          description: "No service ticket"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "no service ticket found"
            status: "Error"

  # add parts to service ticket
  /service_tickets/{ticket_id}/add_part:
    put: 
      tags:
        - ServiceTickets
      summary: "Add a part to a service ticket endpoint"
      description: "Takes a service ticket id and adds a part to the service ticket according to the part id and quantity."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required add a part to a service ticket."
          required: true
          schema:
            $ref: "#/definitions/UpdatePartsTicketPayload"
      responses:
        200:
          descriptions: "Successfully added a part to a service ticket"
          schema:
            $ref: "#/definitions/UpdatePartsTicketResponse"
          examples:
            application/json:
              ticket_id: 1
              part_id: 1
              quantity: 1

        404:
          description: "No service ticket"
          schema: 
            type: "object"
          examples:
            application/json:
            messsage: "no service ticket found"
            status: "Error"
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
      
    required:
      - email
      - password

  # customer login response
  LoginResponse:
    type: "object" 
    properties: 
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"

  # create new customer payload
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      
    required:
      - name
      - email
      - phone
      - password

  # create new customer response
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"

  # get customers list response
...
  GetCustomerByID:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"

  # update customer payload
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
      
    required:
      - name
      - email
      - phone
      - password

  # update customer response
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"

  # delete customer response
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string" 

# Mechanics
  # create new mechanic payload
...
  CreateServiceTicketPayload:
    type: "object"
    properties:
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
      customer_id:
        type: "integer"
      
    required:
      - VIN
      - service_date
      - service_desc
      - customer_id

  # create new service ticket response
  CreateServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
      customer_id:
        type: "integer"

  # get service tickets list response
...
  GetCustomerServiceTickets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        VIN:
          type: "string"
        service_date:
          type: "string"
          format: "date"
        service_desc:
          type: "string"
        customer_id:
          type: "integer"

  # edit mechanics on service ticket payload
  UpdateMechanicsTicketPayload:
    type: "object"
    properties:
      ticket_id:
        type: "integer"
      add_mechanic_ids:
        type: "array"
      remove_mechanic_ids:
        type: "array"
      
    required:
      - ticket_id
      - add_mechanic_ids
      - remove_mechanic_ids

  # edit mechanics on service ticket response
  UpdateMechanicsTicketResponse:
    type: "object"
    properties:
      ticket_id:
        type: "integer"
      add_mechanic_ids:
        type: "array"
      remove_mechanic_ids:
        type: "array"

  # add parts on service ticket payload
  UpdatePartsTicketPayload:
    type: "object"
    properties:
      ticket_id:
        type: "integer"
      part_id:
        type: "integer"
      quantity:
        type: "integer"
      
    required:
      - ticket_id
      - part_id
      - quantity

  # add parts on service ticket response
  UpdatePartsTicketResponse:
    type: "object"
    properties:
      ticket_id:
        type: "integer"
      part_id:
        type: "integer"
      quantity:
        type: "integer"